Move RegridderWeightsCache to imod.util as it is part of the public API
