+++ imod/common/__init__.py
@@ -1,1 +1,2 @@
-from imod.common.utilities.regrid import RegridderType, RegridderWeightsCache
+from imod.common.utilities.regrid import RegridderType
+from imod.util.regrid import RegridderWeightsCache
