+++ imod/mf6/utilities/imod5_converter.py
@@ -4,12 +4,13 @@ import numpy as np
 import pandas as pd
 import xarray as xr
 
-from imod.common.utilities.regrid import RegridderWeightsCache, _regrid_package_data
+from imod.common.utilities.regrid import _regrid_package_data
 from imod.common.utilities.regrid_method_type import RegridMethodType
 from imod.mf6.package import Package
 from imod.typing import GridDataDict, Imod5DataDict
 from imod.typing.grid import full_like
 from imod.util.dims import drop_layer_dim_cap_data
+from imod.util.regrid import RegridderWeightsCache
 
 
 def convert_ibound_to_idomain(
