+++ imod/tests/test_mf6/test_mf6_regrid_simulation.py
@@ -4,7 +4,6 @@ import numpy as np
 import pytest
 
 import imod
-from imod.common.utilities.regrid import RegridderWeightsCache
 from imod.mf6 import VerticesDiscretization
 from imod.mf6.regrid.regrid_schemes import ConstantHeadRegridMethod
 from imod.tests.fixtures.mf6_modelrun_fixture import assert_simulation_can_run
@@ -12,7 +11,7 @@ from imod.tests.fixtures.mf6_small_models_fixture import (
     grid_data_structured,
     grid_data_unstructured,
 )
-from imod.util.regrid import RegridderType
+from imod.util.regrid import RegridderType, RegridderWeightsCache
 
 
 def test_regrid_structured_simulation_to_structured_simulation(
