+++ imod/tests/test_msw/test_grid_data.py
@@ -11,10 +11,10 @@ from numpy import nan
 from numpy.testing import assert_almost_equal, assert_equal
 from pytest_cases import case, parametrize_with_cases
 
-from imod.common.utilities.regrid import RegridderWeightsCache
 from imod.mf6.dis import StructuredDiscretization
 from imod.msw import GridData
 from imod.msw.fixed_format import format_fixed_width
+from imod.util.regrid import RegridderWeightsCache
 from imod.util.spatial import get_total_grid_area
